+++ extension/src/types/time-sheet.ts
@@ -1,4 +1,4 @@
-import { isEqual } from 'lodash'
+import equal from 'fast-deep-equal'
 import { z } from 'zod'
 import { DateOnly, DayOfWeek } from './dates'
 
@@ -69,7 +69,7 @@ export interface DaysFilled {
 export const isTimeSheet = (input: any): input is TimeSheet => TimeSheet.safeParse(input).success
 
 export const isTimeSheetEqual = (first: TimeSheet | null, second: TimeSheet | null) => {
-  return isEqual(first, second)
+  return equal(first, second)
 }
 
 export interface TimeSheetSummary {