+++ extension/tools/webpack/webpack.config.dev.js
@@ -1,0 +1,33 @@
+module.exports = {
+  mode: 'development',
+  entry: {
+    popup: './src/popup/popup.ts',
+    'extension-options': './src/extension-options/extension-options.ts',
+    offscreen: './src/offscreen/offscreen.ts',
+  },
+  module: {
+    rules: require('./webpack.rules'),
+  },
+  output: {
+    filename: '[name].js',
+    chunkFilename: '[name].chunk.js',
+  },
+  plugins: require('./webpack.plugins'),
+  resolve: {
+    extensions: ['.js', '.ts', '.jsx', '.tsx', '.css'],
+    alias: require('./webpack.aliases'),
+  },
+  stats: 'errors-warnings',
+  devtool: 'cheap-module-source-map',
+  devServer: {
+    open: true,
+  },
+  optimization: {
+    splitChunks: {
+      chunks: 'all',
+    },
+  },
+  performance: {
+    hints: false,
+  },
+}